reorganize code, add resource hook catalog
