+++ superset-frontend/src/common/hooks/apiResources/index.ts
@@ -17,31 +17,13 @@
  * under the License.
  */
 
-import React from 'react';
-import { useChartOwnerNames } from 'src/common/hooks/apiResources';
-import ErrorMessageWithStackTrace from 'src/components/ErrorMessage/ErrorMessageWithStackTrace';
-import { SupersetError } from 'src/components/ErrorMessage/types';
+export {
+  useApiResourceFullBody,
+  useApiV1Resource,
+  useResourceTransform,
+} from './apiResources';
 
-interface Props {
-  chartId: string;
-  error?: SupersetError;
-}
-
-/**
- * fetches the chart owners and adds them to the extra data of the error message
- */
-export const ChartErrorMessage: React.FC<Props> = ({
-  chartId,
-  error,
-  ...props
-}) => {
-  const { result: owners } = useChartOwnerNames(chartId);
-
-  // don't mutate props
-  const ownedError = error && {
-    ...error,
-    extra: { ...error.extra, owners },
-  };
-
-  return <ErrorMessageWithStackTrace {...props} error={ownedError} />;
-};
+// A central catalog of API hooks.
+// Add new API hooks here, organized under
+// different files for different resource types.
+export { useChartOwnerNames } from './charts';
